Issue 1192: Can't scroll /stats
HACK HACK HACK
There's a weird select2/browser (IE10,11, Safari, some FireFox)
interaction where select2 will try to ensure the selected option is
visible by scrolling the list ('ensureHighlightVisible').
(related: https://github.com/select2/select2/issues/1408 )
Instead, on some browsers, it ends up scrolling the document, and never
succeeds.  Because this is a bug in select2 *somewhere*, here's a hack
to make it stop being silly.
Opening the list so that it can scroll to whereit wants to be, and then
closing it again because at that point it won't try to scroll anymore.
Currently, this only applies to the Duration option as the others don't
have enough items to be scrollable, although some users report issues
even with these - perhaps on mobile (Safari)?
Note that this also only applies if the pokemon result list makes the
page scrollable in the first place, and if the result list is not hidden
in updateStatMap(), there's also no problem.
The latter may be a more appropriate method of fixing this, as its
content will be updated anyway, and the loading div should appear on
top.
The hacky fix either happens so fast that the browser's renderer can't
keep up with it, or the draw calls never make it out.  Either way, this
is completely invisible in my tests while 'fixing' the issue.
HACK HACK HACK