Issue 1242: Request notification only if needed
This change moves the request for notification permissions to those
functions that actually need it.
In addition, it provides visual feedback via a new div to inform the
user about their current state of notification permissions.
This also fixes repeat notification permissions if the user simply
dismissed the dialog rather than choosing to block the permission.
This introduces an in-head style element.  The page is styled through
app.min.css for which there does not appear to be an app.css to modify
directly.