Long Select2 lists improvements
These are improvements on the Select2 lists that tend to be long, i.e.
the Pokémon Exclusion and Notification lists.
1. They are now allowed to be cleared through the Select2 property
'allowClear: true'.  This places a small 'x' marker in the top-right
corner of the list.
2. They no longer close automatically after selecting a 'mon, using
'closeOnSelect=false', allowing you to select further 'mon.
3. When the list gets too long and forces a scrollbar, and the dropdown
would be displayed above, it would migrate further and further up (Issue
1304).  This is a bug in Select2 that was supposedly fixed in 4.0.3, but
that's what we're using, so no.  The fix is a little dirty: it forces
the dropdown to be re-positioned after it opens (actually after a
timeout as the Select2 open event fires off too soon).
With the above (easy multiple selection) also came the need to not have
the nav bar close when selecting 'mon from the list.  This is taken care
of in app.js by having it ignore clicks/taps on the list (UL/LI
elements), rather than having to set up further events on Select2's
dynamic elements.