Issue 1160: Exportable, importable notification
At the basis of Issue #1160 is that it takes a while to select mon to
exclude/notify.  Another branch makes this easier by allowing to select
multiple easier.  However, another request in Issue #1160 is moving this
to another machine.  Rather than building a fairly involved method of
exporting/importing these lists to/from files, the lists can effectively
be shared using a bookmarklet that changes the appropriate list's value
and calls its change 'event' handler.
The code changes here implement exactly that, adding a 'shareable
bookmarklet' link under those two lists.  Added bonus: when dragged, the
link text is changed to indicate the list and number of mon.